relative perplexity is more direct than inertia
perplexity over variance is ambiguous since either metric can lead to similar ratios, but if perplexity is higher than variance the model is unconfident & if perplexity is lower than variance the model is overfitting. Perplexity over perplexity of the single-most correlated feature is more comparable, and directly informs if a combination of variants leads to more confident predictions.